requirements.txt
streamlit
pymongo
openai
python-dotenv
requests
lxml
watchdog
streamlit